feat: enhance exec handler to decode output buffers and improve error handling
fix: update command building to validate username configuration

feat: optimize system static info retrieval with caching and timeout management

feat: integrate systeminformation for disk size retrieval on Windows

feat: add utility functions for network interface management

feat: update load widget to handle platform-specific load metrics

fix: improve network status widget to handle unknown interface states and errors

test: add unit tests for exec output decoding and command building

test: enhance disk utility tests for new parsing functionality

test: expand load and network status widget tests for edge cases
