+++ setup/dashboard/ws/src/widgets/load.ts
@@ -2,7 +2,11 @@ import os from "os";
 import si from "systeminformation";
 
 export const widgetsLoad = async () => {
-    const loadavg = os.loadavg();
     const processes = await si.processes();
-    return `${loadavg.map(l=>l.toFixed(2)).join(" ")} ${processes.all}`;
+    if (process.platform !== "win32") {
+        const loadavg = os.loadavg();
+        return `${loadavg.map(l => l.toFixed(2)).join(" ")} ${processes.all}`;
+    }
+    const cpuLoad = await si.currentLoad();
+    return `${cpuLoad.currentLoad.toFixed(2)}% ${processes.all}`;
 };
